Refactor Quatt entity and sensor code into smaller platform-focused modules.
- split shared and platform-specific entity implementations
- move large sensor description blocks into dedicated modules
- fix pylint issues found during the refactor
- keep behavior unchanged while improving structure and maintainability
