+++ custom_components/quatt/entity.py
@@ -3,8 +3,6 @@
 from __future__ import annotations
 
 from abc import abstractmethod
-from collections.abc import Mapping
-from copy import deepcopy
 from dataclasses import dataclass
 from datetime import date, datetime
 from decimal import Decimal
@@ -19,16 +17,9 @@ from homeassistant.components.binary_sensor import (
 from homeassistant.components.climate import (
     ClimateEntity,
     ClimateEntityDescription,
-    ClimateEntityFeature,
     HVACMode,
 )
-from homeassistant.components.number import (
-    DEFAULT_MAX_VALUE,
-    DEFAULT_MIN_VALUE,
-    DEFAULT_STEP,
-    NumberEntity,
-    NumberEntityDescription,
-)
+from homeassistant.components.number import NumberEntity, NumberEntityDescription
 from homeassistant.components.select import SelectEntity, SelectEntityDescription
 from homeassistant.components.sensor import (
     SensorDeviceClass,
@@ -36,25 +27,13 @@ from homeassistant.components.sensor import (
     SensorEntityDescription,
 )
 from homeassistant.components.switch import SwitchEntity, SwitchEntityDescription
-from homeassistant.const import PRECISION_TENTHS, UnitOfTemperature
 from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
 from homeassistant.helpers.typing import StateType
 from homeassistant.helpers.update_coordinator import CoordinatorEntity
 import homeassistant.util.dt as dt_util
 
-from .api_remote_cic import QuattCicRemoteApiClient
-from .api_remote_home_battery import QuattHomeBatteryApiClient
-from .const import (
-    ALL_ELECTRIC_SYSTEM,
-    ATTRIBUTION,
-    DOMAIN,
-    DUO_HEATPUMP_SYSTEM,
-    NAME,
-    OPENTHERM_SYSTEM,
-    QuattDeviceKind,
-)
+from .const import ATTRIBUTION, DOMAIN, NAME, QuattDeviceKind
 from .coordinator import QuattDataUpdateCoordinator
-from .coordinator_home_battery import QuattHomeBatteryDataUpdateCoordinator
 from .coordinator_remote_cic import QuattCicRemoteDataUpdateCoordinator
 
 _LOGGER = logging.getLogger(__name__)
@@ -175,7 +154,9 @@ class QuattEntity(CoordinatorEntity[QuattDataUpdateCoordinator]):
 
         self._device_name = device_name
         self._device_id = device_id
-        self._attr_unique_id = f"{self._hub_id}:{device_id}:{unique_id_key or sensor_key}"
+        self._attr_unique_id = (
+            f"{self._hub_id}:{device_id}:{unique_id_key or sensor_key}"
+        )
 
         attach_to_hub = device_kind == QuattDeviceKind.HUB
         is_service_device = device_kind == QuattDeviceKind.SERVICE
@@ -263,49 +244,6 @@ class QuattSensor(QuattEntity, SensorEntity):
         return value
 
 
-class QuattChillSensor(QuattSensor):
-    """Quatt Chill sensor class."""
-
-    @property
-    def native_value(self) -> StateType | date | datetime | Decimal:
-        """Return the native value of the Chill sensor."""
-        chill_index = self._current_chill_index()
-        if chill_index is None:
-            return None
-
-        value = self.coordinator.get_value(
-            self._chill_key_at_index(self.entity_description.key, chill_index)
-        )
-
-        if value is None:
-            return None
-
-        if self.entity_description.device_class == SensorDeviceClass.TIMESTAMP:
-            value = dt_util.parse_datetime(value)
-
-        if not isinstance(value, str):
-            return value
-
-        key = self.entity_description.key.rsplit(".", 1)[-1]
-        if key not in CHILL_SENTENCE_CASE_KEYS:
-            return value
-
-        return _format_chill_enum_value(key, value)
-
-
-class QuattSystemSensor(QuattSensor):
-    """Quatt System Sensor class."""
-
-    @property
-    def extra_state_attributes(self) -> dict[str, bool]:
-        """Expose Quatt feature flags as state attributes."""
-        return {
-            DUO_HEATPUMP_SYSTEM: self.coordinator.heatpump_2_active(),
-            ALL_ELECTRIC_SYSTEM: self.coordinator.all_electric_active(),
-            OPENTHERM_SYSTEM: self.coordinator.is_boiler_opentherm(),
-        }
-
-
 class QuattBinarySensor(QuattEntity, BinarySensorEntity):
     """Quatt BinarySensor class."""
 
@@ -338,20 +276,6 @@ class QuattBinarySensor(QuattEntity, BinarySensorEntity):
     def is_on(self) -> bool:
         """Return true if the binary_sensor is on."""
         return self.coordinator.get_value(self.entity_description.key)
-
-
-class QuattChillBinarySensor(QuattBinarySensor):
-    """Quatt Chill binary sensor class."""
-
-    @property
-    def is_on(self) -> bool:
-        """Return true if the Chill binary sensor is on."""
-        chill_index = self._current_chill_index()
-        if chill_index is None:
-            return False
-        return self.coordinator.get_value(
-            self._chill_key_at_index(self.entity_description.key, chill_index)
-        )
 
 
 class QuattSelect(QuattEntity, SelectEntity):
@@ -426,48 +350,6 @@ class QuattSelect(QuattEntity, SelectEntity):
 
         # Always refresh coordinator data after a successful update
         await self.coordinator.async_request_refresh()
-
-
-class QuattSoundSelect(QuattSelect):
-    """Select entity for Quatt sound level configuration."""
-
-    async def _perform_api_update(self, option: str) -> bool:
-        """Perform paired day/night sound level update."""
-
-        remote_client = self.coordinator.client
-        if not isinstance(remote_client, QuattCicRemoteApiClient):
-            _LOGGER.error(
-                "Cannot update %s: remote client required", self.entity_description.key
-            )
-            return False
-
-        # Get current values for both sound levels
-        day_level = self.coordinator.get_value("dayMaxSoundLevel")
-        night_level = self.coordinator.get_value("nightMaxSoundLevel")
-
-        # Update the value that changed
-        if self.entity_description.key == "dayMaxSoundLevel":
-            day_level = option
-        elif self.entity_description.key == "nightMaxSoundLevel":
-            night_level = option
-
-        # Validate that we have both values
-        if not day_level or not night_level:
-            _LOGGER.error(
-                "Cannot update sound level: missing current values (day=%s, night=%s)",
-                day_level,
-                night_level,
-            )
-            return False
-
-        # Send both values to the API
-        settings = {
-            "dayMaxSoundLevel": day_level,
-            "nightMaxSoundLevel": night_level,
-        }
-
-        _LOGGER.debug("Updating CIC sound levels: %s", settings)
-        return await remote_client.update_cic_settings(settings)
 
 
 class QuattSwitch(QuattEntity, SwitchEntity):
@@ -554,37 +436,6 @@ class QuattSwitch(QuattEntity, SwitchEntity):
 
         # Always refresh coordinator data after a successful update
         await self.coordinator.async_request_refresh()
-
-
-class QuattSettingSwitch(QuattSwitch):
-    """Switch entity for Quatt boolean settings."""
-
-    async def _perform_api_update(self, state: bool) -> bool:
-        """Perform boolean setting update."""
-        remote_client = self.coordinator.client
-        if not isinstance(remote_client, QuattCicRemoteApiClient):
-            _LOGGER.error(
-                "Cannot update %s: remote client required", self.entity_description.key
-            )
-            return False
-
-        # Convert dot notation to nested object structure
-        key_parts = self.entity_description.key.split(".")
-
-        # Build nested dictionary from dot notation
-        settings = {}
-        current = settings
-        for i, part in enumerate(key_parts):
-            if i == len(key_parts) - 1:
-                # Last part - set the actual value
-                current[part] = state
-            else:
-                # Intermediate part - create nested dict
-                current[part] = {}
-                current = current[part]
-
-        _LOGGER.debug("Updating CIC setting: %s", settings)
-        return await remote_client.update_cic_settings(settings)
 
 
 class QuattNumber(QuattEntity, NumberEntity):
@@ -685,108 +536,6 @@ class QuattNumber(QuattEntity, NumberEntity):
         await self.coordinator.async_request_refresh()
 
 
-class QuattHomeBatterySolarCapacityNumber(QuattNumber):
-    """Number entity for the home battery installation's ``solarCapacitykWp``.
-
-    The value is a flat scalar on the installation record (not wrapped in
-    ``{value,minValue,maxValue,increment}``), so min/max/step come from the
-    entity description instead of the coordinator data.
-    """
-
-    @property
-    def native_value(self) -> float | None:
-        """Read the scalar value directly from the coordinator data."""
-        return self.coordinator.get_value(self.entity_description.key)
-
-    @property
-    def native_min_value(self) -> float:
-        """Use the entity description's min value, else HA's default."""
-        value = self.entity_description.native_min_value
-        return value if value is not None else DEFAULT_MIN_VALUE
-
-    @property
-    def native_max_value(self) -> float:
-        """Use the entity description's max value, else HA's default."""
-        value = self.entity_description.native_max_value
-        return value if value is not None else DEFAULT_MAX_VALUE
-
-    @property
-    def native_step(self) -> float | None:
-        """Use the entity description's step, else HA's default."""
-        value = self.entity_description.native_step
-        return value if value is not None else DEFAULT_STEP
-
-    async def _perform_api_update(self, value: float) -> bool:
-        """Send the PATCH to the home battery installation endpoint."""
-        client = self.coordinator.client
-        if not isinstance(client, QuattHomeBatteryApiClient):
-            _LOGGER.error(
-                "Cannot update %s: home battery client required",
-                self.entity_description.key,
-            )
-            return False
-        return await client.update_solar_capacity(value)
-
-    async def async_set_native_value(self, value: float) -> None:
-        """Set the new value via the home battery coordinator."""
-        if not isinstance(self.coordinator, QuattHomeBatteryDataUpdateCoordinator):
-            _LOGGER.error(
-                "Cannot update %s: home battery coordinator required",
-                self.entity_description.key,
-            )
-            raise NotImplementedError(
-                f"Setting {self.entity_description.key} requires a home battery hub"
-            )
-
-        try:
-            success = await self._perform_api_update(value)
-        except Exception as err:
-            _LOGGER.exception("Error updating %s", self.entity_description.key)
-            raise RuntimeError(
-                f"Failed to update {self.entity_description.key}"
-            ) from err
-
-        if not success:
-            raise RuntimeError(f"Failed to update {self.entity_description.key}")
-
-        await self.coordinator.async_request_refresh()
-
-
-class QuattSettingNumber(QuattNumber):
-    """Number entity for Quatt numeric settings stored as {value, minValue, maxValue, increment}."""
-
-    async def _perform_api_update(self, value: float) -> bool:
-        """Perform numeric setting update."""
-        remote_client = self.coordinator.client
-        if not isinstance(remote_client, QuattCicRemoteApiClient):
-            _LOGGER.error(
-                "Cannot update %s: remote client required", self.entity_description.key
-            )
-            return False
-
-        # Preserve integer type when the step is whole-number
-        step = self.native_step
-        if step is not None and float(step).is_integer() and float(value).is_integer():
-            payload_value: int | float = int(value)
-        else:
-            payload_value = value
-
-        # Convert dot notation to nested object structure ending at ".value"
-        key_parts = self.entity_description.key.split(".") + ["value"]
-
-        settings: dict[str, Any] = {}
-        current = settings
-        for i, part in enumerate(key_parts):
-            if i == len(key_parts) - 1:
-                current[part] = payload_value
-            else:
-                current[part] = {}
-                current = current[part]
-
-        _LOGGER.debug("Updating CIC setting: %s", settings)
-        return await remote_client.update_cic_settings(settings)
-
-
 @dataclass(frozen=True)
 class QuattFeatureFlags:
     """Quatt feature flags used an entity."""
@@ -846,288 +595,3 @@ class QuattClimateEntityDescription(ClimateEntityDescription, frozen_or_thawed=T
     quatt_features: QuattFeatureFlags = QuattFeatureFlags()
     quatt_entity_class: type[QuattEntity] = ClimateEntity
     quatt_unique_id_key: str | None = None
-
-
-class QuattChillClimate(QuattEntity, ClimateEntity):
-    """Climate entity for Quatt chill devices."""
-
-    _attr_fan_modes = [fan_mode.display_value for fan_mode in QuattChillFanMode]
-    _attr_hvac_modes = [HVACMode.HEAT, HVACMode.COOL, HVACMode.OFF]
-    _attr_name = None
-    _attr_precision = PRECISION_TENTHS
-    _attr_target_temperature_step = 1.0
-    _attr_supported_features = (
-        ClimateEntityFeature.TARGET_TEMPERATURE | ClimateEntityFeature.FAN_MODE
-    )
-
-    def __init__(
-        self,
-        device_name: str,
-        device_id: str,
-        sensor_key: str,
-        coordinator: QuattDataUpdateCoordinator,
-        entity_description: QuattClimateEntityDescription,
-        device_kind: QuattDeviceKind,
-    ) -> None:
-        """Initialize the chill climate entity."""
-        super().__init__(
-            device_name,
-            device_id,
-            sensor_key,
-            coordinator,
-            device_kind,
-            entity_description.quatt_unique_id_key,
-        )
-        self.entity_description = entity_description
-
-    @property
-    def chill_index(self) -> int | None:
-        """Return the current Chill list index for this climate entity."""
-        return self._current_chill_index()
-
-    def _chill_value(self, key: str) -> Any:
-        """Return a value from the current Chill data."""
-        chill_index = self.chill_index
-        if chill_index is None:
-            return None
-        return self.coordinator.get_value(f"chills.{chill_index}.{key}")
-
-    def _async_set_chill_values(self, updates: Mapping[str, Any]) -> None:
-        """Optimistically update the cached chill values."""
-        chill_index = self.chill_index
-        if chill_index is None:
-            return
-
-        data = deepcopy(self.coordinator.data)
-        current_node = data
-        if isinstance(current_node, dict) and "result" in current_node:
-            current_node = current_node["result"]
-
-        if not isinstance(current_node, dict):
-            return
-
-        chills = current_node.get("chills")
-        if not isinstance(chills, list) or chill_index >= len(chills):
-            return
-
-        chill = chills[chill_index]
-        if not isinstance(chill, dict):
-            return
-
-        for value_path, value in updates.items():
-            node = chill
-            parts = value_path.split(".")
-            for part in parts[:-1]:
-                child = node.setdefault(part, {})
-                if not isinstance(child, dict):
-                    return
-                node = child
-            node[parts[-1]] = value
-
-        self.coordinator.async_set_updated_data(data)
-
-    @staticmethod
-    def _status_to_hvac_mode(status: Any, source: str) -> HVACMode | None:
-        """Convert a Quatt Chill status or mode value to an HVAC mode."""
-        if chill_status := QuattChillStatus.from_api_value(status):
-            return chill_status.hvac_mode
-        if chill_mode := QuattChillMode.from_api_value(status):
-            return chill_mode.hvac_mode
-        if isinstance(status, str):
-            _LOGGER.debug("Unknown Chill %s value: %s", source, status)
-        return None
-
-    @property
-    def hvac_mode(self) -> HVACMode:
-        """Return the current HVAC mode."""
-        status = self._chill_value("status")
-        if hvac_mode := self._status_to_hvac_mode(status, "status"):
-            return hvac_mode
-
-        is_on = self._chill_value("isOn.value")
-        if not is_on:
-            return HVACMode.OFF
-
-        mode = self._chill_value("mode")
-        if hvac_mode := self._status_to_hvac_mode(mode, "mode"):
-            return hvac_mode
-
-        return HVACMode.OFF
-
-    @property
-    def current_temperature(self) -> float | None:
-        """Return the current temperature."""
-        return self._chill_value("ambientTemperature")
-
-    @property
-    def target_temperature(self) -> float | None:
-        """Return the target temperature."""
-        hvac_mode = self.hvac_mode
-        if hvac_mode == HVACMode.OFF:
-            mode = self._chill_value("mode")
-            hvac_mode = self._status_to_hvac_mode(mode, "mode")
-
-        if hvac_mode == HVACMode.COOL:
-            return self._chill_value("coolingTargetTemperature")
-        if hvac_mode == HVACMode.HEAT:
-            return self._chill_value("heatingTargetTemperature")
-        return None
-
-    @property
-    def target_temperature_high(self) -> float | None:
-        """Return the maximum target temperature."""
-        return self._chill_value("maxTargetTemperature")
-
-    @property
-    def target_temperature_low(self) -> float | None:
-        """Return the minimum target temperature."""
-        return self._chill_value("minTargetTemperature")
-
-    @property
-    def min_temp(self) -> float:
-        """Return the minimum temperature."""
-        min_temp = self._chill_value("minTargetTemperature")
-        return float(min_temp) if min_temp is not None else 5.0
-
-    @property
-    def max_temp(self) -> float:
-        """Return the maximum temperature."""
-        max_temp = self._chill_value("maxTargetTemperature")
-        return float(max_temp) if max_temp is not None else 40.0
-
-    @property
-    def temperature_unit(self) -> str:
-        """Return the unit of measurement."""
-        return UnitOfTemperature.CELSIUS
-
-    @property
-    def fan_mode(self) -> str | None:
-        """Return the current fan mode."""
-        fan_mode = self._chill_value("fanMode")
-        if isinstance(fan_mode, str):
-            return _format_chill_enum_value("fanMode", fan_mode)
-        return None
-
-    async def async_set_hvac_mode(self, hvac_mode: HVACMode) -> None:
-        """Set the HVAC mode."""
-        remote_client = self.coordinator.client
-        if not isinstance(remote_client, QuattCicRemoteApiClient):
-            raise NotImplementedError("HVAC mode setting requires remote API")
-
-        chill_uuid = self._chill_value("uuid")
-        if not chill_uuid:
-            raise RuntimeError("Cannot find chill UUID")
-
-        if hvac_mode == HVACMode.OFF:
-            data = {"type": "SET_ON_OFF", "on": False}
-            _LOGGER.debug("Turning off chill %s: %s", chill_uuid, data)
-            success = await remote_client.update_chill_action(chill_uuid, data)
-            if not success:
-                raise RuntimeError(f"Failed to set HVAC mode to {hvac_mode}")
-
-            self._async_set_chill_values(
-                {
-                    "isOn.value": False,
-                    "status": QuattChillStatus.OFF.value,
-                }
-            )
-            return
-
-        if hvac_mode == HVACMode.COOL:
-            chill_mode = QuattChillMode.COOLING
-            chill_status = QuattChillStatus.COOLING
-        elif hvac_mode == HVACMode.HEAT:
-            chill_mode = QuattChillMode.HEATING
-            chill_status = QuattChillStatus.HEATING
-        else:
-            raise ValueError(f"Unsupported HVAC mode: {hvac_mode}")
-
-        if self.hvac_mode == HVACMode.OFF:
-            data = {"type": "SET_ON_OFF", "on": True}
-            _LOGGER.debug("Turning on chill %s: %s", chill_uuid, data)
-            success = await remote_client.update_chill_action(chill_uuid, data)
-            if not success:
-                raise RuntimeError(f"Failed to set HVAC mode to {hvac_mode}")
-            self._async_set_chill_values({"isOn.value": True})
-
-        data = {"type": "SET_MODE", "mode": chill_mode.value}
-        _LOGGER.debug("Setting HVAC mode for chill %s: %s", chill_uuid, data)
-        success = await remote_client.update_chill_action(chill_uuid, data)
-        if not success:
-            raise RuntimeError(f"Failed to set HVAC mode to {hvac_mode}")
-
-        self._async_set_chill_values(
-            {
-                "isOn.value": True,
-                "mode": chill_mode.value,
-                "status": chill_status.value,
-            }
-        )
-
-    async def async_set_fan_mode(self, fan_mode: str) -> None:
-        """Set the fan mode."""
-        remote_client = self.coordinator.client
-        if not isinstance(remote_client, QuattCicRemoteApiClient):
-            raise NotImplementedError("Fan mode setting requires remote API")
-
-        chill_fan_mode = QuattChillFanMode.from_display_value(fan_mode)
-        if chill_fan_mode is None:
-            raise ValueError(f"Unsupported fan mode: {fan_mode}")
-
-        chill_uuid = self._chill_value("uuid")
-        if not chill_uuid:
-            raise RuntimeError("Cannot find chill UUID")
-
-        data = {"type": "SET_FAN_MODE", "fanMode": chill_fan_mode.value}
-        _LOGGER.debug("Setting fan mode for chill %s: %s", chill_uuid, data)
-        success = await remote_client.update_chill_action(chill_uuid, data)
-        if success:
-            self._async_set_chill_values({"fanMode": chill_fan_mode.value})
-        else:
-            raise RuntimeError(f"Failed to set fan mode to {fan_mode}")
-
-    async def async_set_temperature(self, **kwargs) -> None:
-        """Set the target temperature."""
-        remote_client = self.coordinator.client
-        if not isinstance(remote_client, QuattCicRemoteApiClient):
-            raise NotImplementedError("Temperature setting requires remote API")
-
-        temperature = kwargs.get("temperature")
-        if temperature is None:
-            return
-        temperature = int(round(temperature))
-
-        chill_uuid = self._chill_value("uuid")
-        if not chill_uuid:
-            raise RuntimeError("Cannot find chill UUID")
-
-        hvac_mode = self.hvac_mode
-        if hvac_mode == HVACMode.OFF:
-            mode = self._chill_value("mode")
-            hvac_mode = self._status_to_hvac_mode(mode, "mode")
-
-        if hvac_mode == HVACMode.COOL:
-            data = {
-                "type": "SET_COOLING_TARGET_TEMPERATURE",
-                "coolingTargetTemperature": temperature,
-            }
-        elif hvac_mode == HVACMode.HEAT:
-            data = {
-                "type": "SET_HEATING_TARGET_TEMPERATURE",
-                "heatingTargetTemperature": temperature,
-            }
-        else:
-            raise RuntimeError(f"Unknown chill mode: {hvac_mode}")
-
-        _LOGGER.debug("Setting target temperature for chill %s: %s", chill_uuid, data)
-        success = await remote_client.update_chill_action(chill_uuid, data)
-        if success:
-            self._async_set_chill_values(
-                {
-                    "coolingTargetTemperature"
-                    if hvac_mode == HVACMode.COOL
-                    else "heatingTargetTemperature": temperature
-                }
-            )
-        else:
-            raise RuntimeError(f"Failed to set temperature to {temperature}")
